1. AOP를 활용한 예외처리 구현

+++ src/main/java/com/sparta/hanghae99springlv1/service/ReplyService.java
@@ -25,7 +25,7 @@ public class ReplyService {
 
         // 게시글 id로 DB 조회
         Post post = postRepository.findById(postId)
-                .orElseThrow(() -> new IllegalArgumentException("게시글이 존재하지 않습니다."));
+                .orElseThrow(() -> new IllegalArgumentException("게시글을 찾을 수 없습니다."));
 
         // 오청받은 DTO로 DB에 저장할 객체 만들기
         Reply reply = replyRepository.saveAndFlush(new Reply(requestDto, post, user));
@@ -36,39 +36,29 @@ public class ReplyService {
     // 댓글 수정
     @Transactional
     public ReplyResponseDto updateReply(Long replyId, ReplyRequestDto requestDto, User user) {
-        // 사용자 권한 가져와서
-        UserRoleEnum userRoleEnum = user.getRole();
+        // 게시글이 존재하는지 확인
+        Reply reply = replyRepository.findById(replyId)
+                .orElseThrow(() -> new IllegalArgumentException("댓글을 찾을 수 없습니다."));
 
-        // ADMIN 이면 모든 댓글 수정 가능
-        Reply reply;
-        if (userRoleEnum == UserRoleEnum.ADMIN) {
-            reply = replyRepository.findById(replyId)
-                    .orElseThrow(() -> new NullPointerException("댓글이 존재하지 않습니다."));
-        } else {
-            // USER 이면 자기 게시글 수정 가능, 현재 유저가 작성한 게시글이 맞는지 검증
-            reply = replyRepository.findByIdAndUser_Id(replyId, user.getId())
-                    .orElseThrow(() -> new NullPointerException("현재 사용자가 작성한 댓글이 아닙니다."));
+        // 사용자 권한이 USER이고, 자기가 작성한 댓글이 아니라면 오류 발생
+        if (user.getRole() == UserRoleEnum.USER && !reply.getUser().getId().equals(user.getId())) {
+            throw new IllegalArgumentException("현재 사용자가 작성한 댓글이 아닙니다.");
         }
 
         reply.update(requestDto);
-
         return new ReplyResponseDto(reply);
     }
 
     // 댓글 삭제
     @Transactional
     public Message deleteReply(Long replyId, User user) {
-        // 사용자 권한 가져와서
-        UserRoleEnum userRoleEnum = user.getRole();
+        // 게시글이 존재하는지 확인
+        Reply reply = replyRepository.findById(replyId)
+                .orElseThrow(() -> new IllegalArgumentException("댓글을 찾을 수 없습니다."));
 
-        // ADMIN 이면 모든 댓글 삭제 가능
-        if (userRoleEnum == UserRoleEnum.ADMIN) {
-            Reply reply = replyRepository.findById(replyId)
-                    .orElseThrow(() -> new NullPointerException("댓글이 존재하지 않습니다."));
-        } else {
-            // USER 이면 자기 게시글 삭제 가능, 현재 유저가 작성한 게시글이 맞는지 검증
-            Reply reply = replyRepository.findByIdAndUser_Id(replyId, user.getId())
-                    .orElseThrow(() -> new NullPointerException("현재 사용자가 작성한 댓글이 아닙니다."));
+        // 사용자 권한이 USER이고, 자기가 작성한 댓글이 아니라면 오류 발생
+        if (user.getRole() == UserRoleEnum.USER && !reply.getUser().getId().equals(user.getId())) {
+            throw new IllegalArgumentException("현재 사용자가 작성한 댓글이 아닙니다.");
         }
 
         replyRepository.deleteById(replyId);
